debug pm manager

+++ mad_cdev/vm_cdev.c
@@ -44,7 +44,7 @@ uint64_t x97_address_translate(dev_pmap_t *pmap, void *va) {
     // printf("[x97_address_translate] done, page pa: %lx\n", *pte);
     if (*pte != 0) {
         if (pmap->mode == EXCLUSIVE && (*pte < pmap->mmu_ops->pa_min || *pte >= pmap->mmu_ops->pa_max)) {
-            printf("PDE page status: refcount %u, flags %u", PHYS_TO_VM_PAGE(DMAP_TO_PHYS(pte))->ref_count, PHYS_TO_VM_PAGE(DMAP_TO_PHYS(pte))->flags);
+            printf("PDE page status: refcount %u, flags %u", PHYS_TO_VM_PAGE(DMAP_TO_PHYS((uintptr_t) pte))->ref_count, PHYS_TO_VM_PAGE(DMAP_TO_PHYS((uintptr_t) pte))->flags);
             printf("%s %d: translate va %lx - pa %lx\n",__func__, __LINE__, (uintptr_t) va, *pte);
             return 0;
         }
